Simon-game-website
Rules to play the game : 1- To start the game one should press any key from the keyboard. At the beginning of the game, one of the button lights up randomly producing simultaneously a sound associated to the button. 2- The player has to press the same button. 3- Next, the Simon turns a second one, again randomly. 4- The player has to reproduce this chain of light using his memory. 5- And so on... In each round a new button is added to the series and the game becomes all the more difficult as the player's memory is put to the test. 6- If the player doesn't make any mistake, the game goes on, so it is an endless game!